ai-platform-unified/tutorials/unofficial/pytorch-image-classification-custom-train-container-multi-replica/trainer/requirements.txt
torch==1.8.1
torchvision==0.9.1
google-cloud-storage